chore: split `scanner.rs` into different files
