--- crates/rolldown/src/bundler/ast_scanner/scanner.rs
@@ -1,598 +1,0 @@
-use std::{borrow::Cow, sync::Arc};
-
-use index_vec::IndexVec;
-use oxc::{
-  ast::{
-    ast::{
-      ExportAllDeclaration, ExportDefaultDeclaration, ExportNamedDeclaration, IdentifierReference,
-      ImportDeclaration, MemberExpression, ModuleDeclaration, Program,
-    },
-    Visit,
-  },
-  semantic::SymbolId,
-  span::{Atom, Span},
-};
-use rolldown_common::{
-  representative_name, ExportsKind, FilePath, ImportKind, ImportRecordId, LocalExport, ModuleId,
-  ModuleType, NamedImport, RawImportRecord, Specifier, StmtInfo, StmtInfos, SymbolRef,
-};
-use rolldown_error::BuildError;
-use rolldown_oxc::{BindingIdentifierExt, BindingPatternExt};
-use rustc_hash::{FxHashMap, FxHashSet};
-
-use crate::bundler::utils::{ast_scope::AstScope, ast_symbol::AstSymbol};
-
-// Probably we should generate this using macros.
-static SIDE_EFFECT_FREE_MEMBER_EXPR_2: once_cell::sync::Lazy<
-  FxHashSet<(Cow<'static, Atom>, Cow<'static, Atom>)>,
-> = once_cell::sync::Lazy::new(|| {
-  [
-    ("Object", "create"),
-    ("Object", "defineProperty"),
-    ("Object", "getOwnPropertyDescriptor"),
-    ("Object", "getPrototypeOf"),
-    ("Object", "getOwnPropertyNames"),
-  ]
-  .into_iter()
-  .map(|(obj, prop)| (Cow::Owned(obj.into()), Cow::Owned(prop.into())))
-  .collect()
-});
-
-// hyf0: clippy::type_complexity: This is only a temporary solution.
-#[allow(clippy::type_complexity)]
-static SIDE_EFFECT_FREE_MEMBER_EXPR_3: once_cell::sync::Lazy<
-  FxHashSet<(Cow<'static, Atom>, Cow<'static, Atom>, Cow<'static, Atom>)>,
-> = once_cell::sync::Lazy::new(|| {
-  [("Object", "prototype", "hasOwnProperty"), ("Object", "prototype", "constructor")]
-    .into_iter()
-    .map(|(obj, obj_mid, prop)| {
-      (Cow::Owned(obj.into()), Cow::Owned(obj_mid.into()), Cow::Owned(prop.into()))
-    })
-    .collect()
-});
-
-#[derive(Debug, Default)]
-pub struct ScanResult {
-  pub repr_name: String,
-  pub named_imports: FxHashMap<SymbolId, NamedImport>,
-  pub named_exports: FxHashMap<Atom, LocalExport>,
-  pub stmt_infos: StmtInfos,
-  pub import_records: IndexVec<ImportRecordId, RawImportRecord>,
-  pub star_exports: Vec<ImportRecordId>,
-  pub export_default_symbol_id: Option<SymbolId>,
-  pub imports: FxHashMap<Span, ImportRecordId>,
-  pub exports_kind: ExportsKind,
-  pub warnings: Vec<BuildError>,
-}
-
-pub struct AstScanner<'a> {
-  idx: ModuleId,
-  source: &'a Arc<str>,
-  module_type: ModuleType,
-  file_path: &'a FilePath,
-  scope: &'a AstScope,
-  symbol_table: &'a mut AstSymbol,
-  current_stmt_info: StmtInfo,
-  result: ScanResult,
-  esm_export_keyword: Option<Span>,
-  esm_import_keyword: Option<Span>,
-  pub namespace_symbol: SymbolRef,
-  used_exports_ref: bool,
-  used_module_ref: bool,
-}
-
-impl<'ast> AstScanner<'ast> {
-  pub fn new(
-    idx: ModuleId,
-    scope: &'ast AstScope,
-    symbol_table: &'ast mut AstSymbol,
-    repr_name: String,
-    module_type: ModuleType,
-    source: &'ast Arc<str>,
-    file_path: &'ast FilePath,
-  ) -> Self {
-    let mut result = ScanResult::default();
-    let name = format!("{repr_name}_ns");
-    let namespace_ref: SymbolRef =
-      (idx, symbol_table.create_symbol(name.into(), scope.root_scope_id())).into();
-    result.repr_name = repr_name;
-    // The first StmtInfo is to represent the namespace binding.
-    result
-      .stmt_infos
-      .add_stmt_info(StmtInfo { declared_symbols: vec![namespace_ref], ..Default::default() });
-    Self {
-      idx,
-      scope,
-      symbol_table,
-      current_stmt_info: StmtInfo::default(),
-      result,
-      esm_export_keyword: None,
-      esm_import_keyword: None,
-      module_type,
-      namespace_symbol: namespace_ref,
-      used_exports_ref: false,
-      used_module_ref: false,
-      source,
-      file_path,
-    }
-  }
-
-  pub fn scan(mut self, program: &Program<'ast>) -> ScanResult {
-    self.visit_program(program);
-    let mut exports_kind = ExportsKind::None;
-
-    if self.esm_export_keyword.is_some() {
-      exports_kind = ExportsKind::Esm;
-    } else if self.used_exports_ref || self.used_module_ref {
-      exports_kind = ExportsKind::CommonJs;
-    } else {
-      // TODO(hyf0): need review this. Why `ModuleType` doesn't have higher priority?
-      match self.module_type {
-        ModuleType::CJS | ModuleType::CjsPackageJson => {
-          exports_kind = ExportsKind::CommonJs;
-        }
-        ModuleType::EsmMjs | ModuleType::EsmPackageJson => {
-          exports_kind = ExportsKind::Esm;
-        }
-        ModuleType::Unknown => {
-          if self.esm_import_keyword.is_some() {
-            exports_kind = ExportsKind::Esm;
-          }
-        }
-      }
-    }
-
-    self.result.exports_kind = exports_kind;
-    self.result
-  }
-
-  fn is_unresolved_reference(&self, ident_ref: &IdentifierReference) -> bool {
-    self.scope.is_unresolved(ident_ref.reference_id.get().unwrap())
-  }
-
-  fn set_esm_export_keyword(&mut self, span: Span) {
-    self.esm_export_keyword.get_or_insert(span);
-  }
-
-  fn add_declared_id(&mut self, id: SymbolId) {
-    self.current_stmt_info.declared_symbols.push((self.idx, id).into());
-  }
-
-  fn get_root_binding(&self, name: &Atom) -> SymbolId {
-    self.scope.get_root_binding(name).expect("must have")
-  }
-
-  fn add_import_record(&mut self, module_request: &Atom, kind: ImportKind) -> ImportRecordId {
-    let importee_repr = representative_name(module_request);
-    let namespace_ref: SymbolRef = (
-      self.idx,
-      self
-        .symbol_table
-        .create_symbol(format!("import_{importee_repr}").into(), self.scope.root_scope_id()),
-    )
-      .into();
-    self.add_declared_id(namespace_ref.symbol);
-    let rec = RawImportRecord::new(module_request.clone(), kind, namespace_ref);
-    self.result.import_records.push(rec)
-  }
-
-  fn add_named_import(&mut self, local: SymbolId, imported: &Atom, record_id: ImportRecordId) {
-    self.result.named_imports.insert(
-      local,
-      NamedImport {
-        imported: imported.clone().into(),
-        imported_as: (self.idx, local).into(),
-        record_id,
-      },
-    );
-  }
-
-  fn add_star_import(&mut self, local: SymbolId, record_id: ImportRecordId) {
-    self.result.named_imports.insert(
-      local,
-      NamedImport { imported: Specifier::Star, imported_as: (self.idx, local).into(), record_id },
-    );
-  }
-
-  fn add_local_export(&mut self, export_name: &Atom, local: SymbolId) {
-    self
-      .result
-      .named_exports
-      .insert(export_name.clone(), LocalExport { referenced: (self.idx, local).into() });
-  }
-
-  fn add_local_default_export(&mut self, local: SymbolId) {
-    self.result.export_default_symbol_id = Some(local);
-    self
-      .result
-      .named_exports
-      .insert(Atom::new_inline("default"), LocalExport { referenced: (self.idx, local).into() });
-  }
-
-  fn add_re_export(&mut self, export_name: &Atom, imported: &Atom, record_id: ImportRecordId) {
-    let generated_imported_as_ref = (
-      self.idx,
-      self.symbol_table.create_symbol(
-        if export_name.as_str() == "default" {
-          let importee_repr =
-            representative_name(&self.result.import_records[record_id].module_request);
-          format!("{importee_repr}_default").into()
-        } else {
-          export_name.clone()
-        },
-        self.scope.root_scope_id(),
-      ),
-    )
-      .into();
-    self.current_stmt_info.declared_symbols.push(generated_imported_as_ref);
-    let name_import = NamedImport {
-      imported: imported.clone().into(),
-      imported_as: generated_imported_as_ref,
-      record_id,
-    };
-    self.result.named_imports.insert(generated_imported_as_ref.symbol, name_import);
-    self
-      .result
-      .named_exports
-      .insert(export_name.clone(), LocalExport { referenced: generated_imported_as_ref });
-  }
-
-  fn add_star_re_export(&mut self, export_name: &Atom, record_id: ImportRecordId) {
-    let generated_imported_as_ref =
-      (self.idx, self.symbol_table.create_symbol(export_name.clone(), self.scope.root_scope_id()))
-        .into();
-    self.current_stmt_info.declared_symbols.push(generated_imported_as_ref);
-    let name_import =
-      NamedImport { imported: Specifier::Star, imported_as: generated_imported_as_ref, record_id };
-    self.result.named_imports.insert(generated_imported_as_ref.symbol, name_import);
-    self
-      .result
-      .named_exports
-      .insert(export_name.clone(), LocalExport { referenced: generated_imported_as_ref });
-  }
-
-  fn scan_export_all_decl(&mut self, decl: &ExportAllDeclaration) {
-    let id = self.add_import_record(&decl.source.value, ImportKind::Import);
-    if let Some(exported) = &decl.exported {
-      // export * as ns from '...'
-      self.add_star_re_export(exported.name(), id);
-    } else {
-      // export * from '...'
-      self.result.star_exports.push(id);
-    }
-    self.result.imports.insert(decl.span, id);
-  }
-
-  fn scan_export_named_decl(&mut self, decl: &ExportNamedDeclaration) {
-    if let Some(source) = &decl.source {
-      let record_id = self.add_import_record(&source.value, ImportKind::Import);
-      decl.specifiers.iter().for_each(|spec| {
-        self.add_re_export(spec.exported.name(), spec.local.name(), record_id);
-        self.result.imports.insert(decl.span, record_id);
-      });
-    } else {
-      decl.specifiers.iter().for_each(|spec| {
-        self.add_local_export(spec.exported.name(), self.get_root_binding(spec.local.name()));
-      });
-      if let Some(decl) = decl.declaration.as_ref() {
-        match decl {
-          oxc::ast::ast::Declaration::VariableDeclaration(var_decl) => {
-            var_decl.declarations.iter().for_each(|decl| {
-              decl.id.binding_identifiers().into_iter().for_each(|id| {
-                self.result.named_exports.insert(
-                  id.name.clone(),
-                  LocalExport { referenced: (self.idx, id.expect_symbol_id()).into() },
-                );
-              });
-            });
-          }
-          oxc::ast::ast::Declaration::FunctionDeclaration(fn_decl) => {
-            let id = fn_decl.id.as_ref().unwrap();
-            self.add_local_export(&id.name, id.expect_symbol_id());
-          }
-          oxc::ast::ast::Declaration::ClassDeclaration(cls_decl) => {
-            let id = cls_decl.id.as_ref().unwrap();
-            self.add_local_export(&id.name, id.expect_symbol_id());
-          }
-          _ => unreachable!("doesn't support ts now"),
-        }
-      }
-    }
-  }
-
-  // If the reference is a global variable, `None` will be returned.
-  fn resolve_symbol_from_reference(&self, id_ref: &IdentifierReference) -> Option<SymbolId> {
-    let ref_id = id_ref.reference_id.get().expect("must have reference id");
-    self.scope.symbol_id_for(ref_id)
-  }
-  fn scan_export_default_decl(&mut self, decl: &ExportDefaultDeclaration) {
-    let local = match &decl.declaration {
-      oxc::ast::ast::ExportDefaultDeclarationKind::Expression(exp) => match exp {
-        oxc::ast::ast::Expression::Identifier(id_ref) => self.resolve_symbol_from_reference(id_ref),
-        _ => None,
-      },
-      oxc::ast::ast::ExportDefaultDeclarationKind::FunctionDeclaration(fn_decl) => {
-        fn_decl.id.as_ref().map(rolldown_oxc::BindingIdentifierExt::expect_symbol_id)
-      }
-      oxc::ast::ast::ExportDefaultDeclarationKind::ClassDeclaration(cls_decl) => {
-        cls_decl.id.as_ref().map(rolldown_oxc::BindingIdentifierExt::expect_symbol_id)
-      }
-      _ => unreachable!(),
-    };
-
-    let local = local.unwrap_or_else(|| {
-      // For patterns like `export default [expression]`, we need to create
-      // a facade Symbol to represent it.
-      // Notice: Patterns don't include `export default [identifier]`
-      let sym_id = self.symbol_table.create_symbol(
-        Atom::from(format!("{}_default", self.result.repr_name)),
-        self.scope.root_scope_id(),
-      );
-      self.add_declared_id(sym_id);
-      sym_id
-    });
-    self.add_local_default_export(local);
-  }
-
-  fn scan_import_decl(&mut self, decl: &ImportDeclaration) {
-    let id = self.add_import_record(&decl.source.value, ImportKind::Import);
-    self.result.imports.insert(decl.span, id);
-    let Some(specifiers) = &decl.specifiers else { return };
-    specifiers.iter().for_each(|spec| match spec {
-      oxc::ast::ast::ImportDeclarationSpecifier::ImportSpecifier(spec) => {
-        let sym = spec.local.expect_symbol_id();
-        self.add_named_import(sym, spec.imported.name(), id);
-      }
-      oxc::ast::ast::ImportDeclarationSpecifier::ImportDefaultSpecifier(spec) => {
-        self.add_named_import(spec.local.expect_symbol_id(), &Atom::new_inline("default"), id);
-      }
-      oxc::ast::ast::ImportDeclarationSpecifier::ImportNamespaceSpecifier(spec) => {
-        self.add_star_import(spec.local.expect_symbol_id(), id);
-      }
-    });
-  }
-  fn scan_module_decl(&mut self, decl: &ModuleDeclaration) {
-    match decl {
-      oxc::ast::ast::ModuleDeclaration::ImportDeclaration(decl) => {
-        // TODO: this should be the span of `import` keyword, while now it is the span of the whole import declaration.
-        self.esm_import_keyword.get_or_insert(decl.span);
-        self.scan_import_decl(decl);
-      }
-      oxc::ast::ast::ModuleDeclaration::ExportAllDeclaration(decl) => {
-        self.set_esm_export_keyword(decl.span);
-        self.scan_export_all_decl(decl);
-      }
-      oxc::ast::ast::ModuleDeclaration::ExportNamedDeclaration(decl) => {
-        self.set_esm_export_keyword(decl.span);
-        self.scan_export_named_decl(decl);
-      }
-      oxc::ast::ast::ModuleDeclaration::ExportDefaultDeclaration(decl) => {
-        self.set_esm_export_keyword(decl.span);
-        self.scan_export_default_decl(decl);
-      }
-      _ => {}
-    }
-  }
-
-  pub fn add_referenced_symbol(&mut self, id: SymbolId) {
-    self.current_stmt_info.referenced_symbols.push((self.idx, id).into());
-  }
-
-  fn is_top_level(&self, symbol_id: SymbolId) -> bool {
-    self.scope.root_scope_id() == self.symbol_table.scope_id_for(symbol_id)
-  }
-}
-
-impl<'ast> AstScanner<'ast> {
-  fn visit_top_level_stmt(&mut self, stmt: &oxc::ast::ast::Statement<'ast>) {
-    self.current_stmt_info.side_effect =
-      SideEffectDetector { scope: self.scope }.detect_side_effect_of_stmt(stmt);
-    self.visit_statement(stmt);
-  }
-}
-
-impl<'ast> Visit<'ast> for AstScanner<'ast> {
-  #[tracing::instrument(skip_all)]
-  fn visit_program(&mut self, program: &oxc::ast::ast::Program<'ast>) {
-    for (idx, stmt) in program.body.iter().enumerate() {
-      self.current_stmt_info.stmt_idx = Some(idx);
-      self.visit_top_level_stmt(stmt);
-      self.result.stmt_infos.add_stmt_info(std::mem::take(&mut self.current_stmt_info));
-    }
-  }
-
-  fn visit_binding_identifier(&mut self, ident: &oxc::ast::ast::BindingIdentifier) {
-    let symbol_id = ident.symbol_id.get().unwrap();
-    if self.is_top_level(symbol_id) {
-      self.add_declared_id(symbol_id);
-    }
-  }
-
-  fn visit_identifier_reference(&mut self, ident: &IdentifierReference) {
-    let symbol_id = self.resolve_symbol_from_reference(ident);
-    match symbol_id {
-      Some(symbol_id) if self.is_top_level(symbol_id) => {
-        self.add_referenced_symbol(symbol_id);
-      }
-      None => {
-        if ident.name == "module" {
-          self.used_module_ref = true;
-        }
-        if ident.name == "exports" {
-          self.used_exports_ref = true;
-        }
-        if ident.name == "eval" {
-          self.result.warnings.push(
-            BuildError::unsupported_eval(
-              self.file_path.to_string(),
-              Arc::clone(self.source),
-              ident.span,
-            )
-            .with_severity_warning(),
-          );
-        }
-      }
-      _ => {}
-    }
-  }
-
-  fn visit_statement(&mut self, stmt: &oxc::ast::ast::Statement<'ast>) {
-    if let oxc::ast::ast::Statement::ModuleDeclaration(decl) = stmt {
-      self.scan_module_decl(decl.0);
-    }
-    self.visit_statement_match(stmt);
-  }
-
-  fn visit_import_expression(&mut self, expr: &oxc::ast::ast::ImportExpression<'ast>) {
-    if let oxc::ast::ast::Expression::StringLiteral(request) = &expr.source {
-      let id = self.add_import_record(&request.value, ImportKind::DynamicImport);
-      self.result.imports.insert(expr.span, id);
-    }
-  }
-
-  fn visit_call_expression(&mut self, expr: &oxc::ast::ast::CallExpression<'ast>) {
-    match &expr.callee {
-      oxc::ast::ast::Expression::Identifier(ident)
-        if ident.name == "require" && self.is_unresolved_reference(ident) =>
-      {
-        if let Some(oxc::ast::ast::Argument::Expression(
-          oxc::ast::ast::Expression::StringLiteral(request),
-        )) = &expr.arguments.get(0)
-        {
-          let id = self.add_import_record(&request.value, ImportKind::Require);
-          self.result.imports.insert(expr.span, id);
-        }
-      }
-      _ => {}
-    }
-    for arg in &expr.arguments {
-      self.visit_argument(arg);
-    }
-    self.visit_expression(&expr.callee);
-  }
-}
-
-struct SideEffectDetector<'a> {
-  scope: &'a AstScope,
-}
-
-impl<'a> SideEffectDetector<'a> {
-  fn is_unresolved_reference(&self, ident_ref: &IdentifierReference) -> bool {
-    self.scope.is_unresolved(ident_ref.reference_id.get().unwrap())
-  }
-
-  fn detect_side_effect_of_class(&self, cls: &oxc::ast::ast::Class) -> bool {
-    use oxc::ast::ast::ClassElement;
-    cls.body.body.iter().any(|elm| match elm {
-      ClassElement::StaticBlock(static_block) => {
-        static_block.body.iter().any(|stmt| self.detect_side_effect_of_stmt(stmt))
-      }
-      ClassElement::MethodDefinition(_) => false,
-      ClassElement::PropertyDefinition(def) => {
-        (match &def.key {
-          oxc::ast::ast::PropertyKey::Identifier(_)
-          | oxc::ast::ast::PropertyKey::PrivateIdentifier(_) => false,
-          oxc::ast::ast::PropertyKey::Expression(expr) => self.detect_side_effect_of_expr(expr),
-        } || def.value.as_ref().is_some_and(|init| self.detect_side_effect_of_expr(init)))
-      }
-      ClassElement::AccessorProperty(def) => {
-        (match &def.key {
-          oxc::ast::ast::PropertyKey::Identifier(_)
-          | oxc::ast::ast::PropertyKey::PrivateIdentifier(_) => false,
-          oxc::ast::ast::PropertyKey::Expression(expr) => self.detect_side_effect_of_expr(expr),
-        } || def.value.as_ref().is_some_and(|init| self.detect_side_effect_of_expr(init)))
-      }
-      ClassElement::TSAbstractMethodDefinition(_)
-      | ClassElement::TSAbstractPropertyDefinition(_)
-      | ClassElement::TSIndexSignature(_) => unreachable!("ts should be transpiled"),
-    })
-  }
-
-  fn detect_side_effect_of_member_expr(expr: &oxc::ast::ast::MemberExpression) -> bool {
-    // MemberExpression is considered having side effect by default, unless it's some builtin global variable.
-    let MemberExpression::StaticMemberExpression(expr) = expr else {
-      return true;
-    };
-    let prop = &expr.property.name;
-    match &expr.object {
-      oxc::ast::ast::Expression::Identifier(ident) => {
-        let object_ident = &ident.name;
-        !SIDE_EFFECT_FREE_MEMBER_EXPR_2
-          .contains(&(Cow::Borrowed(object_ident), Cow::Borrowed(prop)))
-      }
-      oxc::ast::ast::Expression::MemberExpression(mem_expr) => {
-        let MemberExpression::StaticMemberExpression(mem_expr) = &**mem_expr else {
-          return true;
-        };
-        let mid_prop = &mem_expr.property.name;
-        let oxc::ast::ast::Expression::Identifier(obj_ident) = &mem_expr.object else {
-          return true;
-        };
-        !SIDE_EFFECT_FREE_MEMBER_EXPR_3.contains(&(
-          Cow::Borrowed(&obj_ident.name),
-          Cow::Borrowed(mid_prop),
-          Cow::Borrowed(prop),
-        ))
-      }
-      _ => true,
-    }
-  }
-
-  fn detect_side_effect_of_expr(&self, expr: &oxc::ast::ast::Expression) -> bool {
-    use oxc::ast::ast::Expression;
-    match expr {
-      Expression::BooleanLiteral(_)
-      | Expression::NullLiteral(_)
-      | Expression::NumberLiteral(_)
-      | Expression::BigintLiteral(_)
-      | Expression::RegExpLiteral(_)
-      | Expression::FunctionExpression(_)
-      | Expression::ArrowExpression(_)
-      | Expression::StringLiteral(_) => false,
-      Expression::MemberExpression(mem_expr) => Self::detect_side_effect_of_member_expr(mem_expr),
-      Expression::ClassExpression(cls) => self.detect_side_effect_of_class(cls),
-      // Accessing global variables considered as side effect.
-      Expression::Identifier(ident) => self.is_unresolved_reference(ident),
-      _ => true,
-    }
-  }
-
-  pub fn detect_side_effect_of_stmt(&self, stmt: &oxc::ast::ast::Statement) -> bool {
-    use oxc::ast::ast::{Declaration, Statement};
-    match stmt {
-      Statement::Declaration(decl) => match decl {
-        Declaration::VariableDeclaration(var_decl) => var_decl
-          .declarations
-          .iter()
-          .any(|decl| decl.init.as_ref().is_some_and(|init| self.detect_side_effect_of_expr(init))),
-        Declaration::FunctionDeclaration(_) => false,
-        Declaration::ClassDeclaration(cls_decl) => self.detect_side_effect_of_class(cls_decl),
-        Declaration::UsingDeclaration(_) => todo!(),
-        Declaration::TSTypeAliasDeclaration(_)
-        | Declaration::TSInterfaceDeclaration(_)
-        | Declaration::TSEnumDeclaration(_)
-        | Declaration::TSModuleDeclaration(_)
-        | Declaration::TSImportEqualsDeclaration(_) => unreachable!("ts should be transpiled"),
-      },
-      Statement::ExpressionStatement(expr) => self.detect_side_effect_of_expr(&expr.expression),
-      Statement::BlockStatement(_)
-      | Statement::BreakStatement(_)
-      | Statement::DebuggerStatement(_)
-      | Statement::DoWhileStatement(_)
-      | Statement::EmptyStatement(_)
-      | Statement::ForInStatement(_)
-      | Statement::ForOfStatement(_)
-      | Statement::ForStatement(_)
-      | Statement::IfStatement(_)
-      | Statement::LabeledStatement(_)
-      | Statement::ReturnStatement(_)
-      | Statement::SwitchStatement(_)
-      | Statement::ThrowStatement(_)
-      | Statement::TryStatement(_)
-      | Statement::WhileStatement(_)
-      | Statement::WithStatement(_)
-      | Statement::ModuleDeclaration(_)
-      | Statement::ContinueStatement(_) => true,
-    }
-  }
-}
